# optimize: reduce syscalls using a buffered reader.