Previously, each noise message read would make two syscalls:

1. one to read the length prefix.
2. one to read the encrypted payload.

This patch adds `bufio.Reader` mediation to cushion syscalls, and significantly enhaces throughput in read-dominated connections, such as file transfers (e.g. bitswap, graphsync transfers).

Ideally, `go-buffer-pool` would support `bufio.Reader` backed by the buffer pool, but we'd need to clone the code for `bufio.Reader`, because one cannot pass the buffer.

## before

```
goos: darwin
goarch: amd64
pkg: github.com/libp2p/go-libp2p-noise
BenchmarkTransfer1MB/readBuffer_>_encrypted_message-16         	     648	   1779676 ns/op	 589906535 bytes/sec	    5026 B/op	     146 allocs/op
BenchmarkTransfer1MB/readBuffer_>_encrypted_message-16         	     636	   1798623 ns/op	 583568598 bytes/sec	    5248 B/op	     148 allocs/op
BenchmarkTransfer1MB/readBuffer_>_encrypted_message-16         	     640	   1809091 ns/op	 580192105 bytes/sec	    5026 B/op	     147 allocs/op
BenchmarkTransfer1MB/readBuffer_>_encrypted_message-16         	     631	   1787210 ns/op	 587280284 bytes/sec	    5001 B/op	     148 allocs/op
BenchmarkTransfer1MB/readBuffer_>_encrypted_message-16         	     663	   1767220 ns/op	 593901061 bytes/sec	    4965 B/op	     145 allocs/op
BenchmarkTransfer1MB/readBuffer_<_decrypted_plaintext-16       	     634	   1800482 ns/op	 583012323 bytes/sec	    4964 B/op	     148 allocs/op
BenchmarkTransfer1MB/readBuffer_<_decrypted_plaintext-16       	     664	   1772233 ns/op	 592238568 bytes/sec	    5058 B/op	     144 allocs/op
BenchmarkTransfer1MB/readBuffer_<_decrypted_plaintext-16       	     639	   1812040 ns/op	 579248566 bytes/sec	    5087 B/op	     147 allocs/op
BenchmarkTransfer1MB/readBuffer_<_decrypted_plaintext-16       	     640	   1781963 ns/op	 589026946 bytes/sec	    5378 B/op	     147 allocs/op
BenchmarkTransfer1MB/readBuffer_<_decrypted_plaintext-16       	     649	   1778060 ns/op	 590289980 bytes/sec	    5065 B/op	     146 allocs/op
BenchmarkTransfer100MB/readBuffer_<_decrypted_plaintext-16     	       4	 288196187 ns/op	 363847866 bytes/sec	 1073562 B/op	   65747 allocs/op
BenchmarkTransfer100MB/readBuffer_<_decrypted_plaintext-16     	       4	 290956831 ns/op	 360395678 bytes/sec	 1067716 B/op	   65736 allocs/op
BenchmarkTransfer100MB/readBuffer_<_decrypted_plaintext-16     	       4	 291216402 ns/op	 360074018 bytes/sec	 1071250 B/op	   65743 allocs/op
BenchmarkTransfer100MB/readBuffer_<_decrypted_plaintext-16     	       4	 288129666 ns/op	 363936245 bytes/sec	 1073332 B/op	   65745 allocs/op
BenchmarkTransfer100MB/readBuffer_<_decrypted_plaintext-16     	       4	 288890120 ns/op	 362972984 bytes/sec	 1074554 B/op	   65747 allocs/op
BenchmarkTransfer100MB/readBuffer_>_encrypted_message-16       	       4	 280595599 ns/op	 373704041 bytes/sec	 1061200 B/op	   65730 allocs/op
BenchmarkTransfer100MB/readBuffer_>_encrypted_message-16       	       4	 282260158 ns/op	 371499446 bytes/sec	 1062314 B/op	   65733 allocs/op
BenchmarkTransfer100MB/readBuffer_>_encrypted_message-16       	       4	 281626426 ns/op	 372334890 bytes/sec	 1060226 B/op	   65732 allocs/op
BenchmarkTransfer100MB/readBuffer_>_encrypted_message-16       	       4	 281607457 ns/op	 372358170 bytes/sec	 1060170 B/op	   65732 allocs/op
BenchmarkTransfer100MB/readBuffer_>_encrypted_message-16       	       4	 281233093 ns/op	 372860965 bytes/sec	 1061282 B/op	   65732 allocs/op
BenchmarkTransfer500Mb/readBuffer_>_encrypted_message-16       	       1	1739228990 ns/op	 301450753 bytes/sec	 8209344 B/op	  512048 allocs/op
BenchmarkTransfer500Mb/readBuffer_>_encrypted_message-16       	       1	1746755706 ns/op	 300151483 bytes/sec	 8208200 B/op	  512035 allocs/op
BenchmarkTransfer500Mb/readBuffer_>_encrypted_message-16       	       1	1739862793 ns/op	 301339364 bytes/sec	 8204104 B/op	  512034 allocs/op
BenchmarkTransfer500Mb/readBuffer_>_encrypted_message-16       	       1	1726923628 ns/op	 303598495 bytes/sec	 8208232 B/op	  512036 allocs/op
BenchmarkTransfer500Mb/readBuffer_>_encrypted_message-16       	       1	1758041663 ns/op	 298223437 bytes/sec	 8213712 B/op	  512057 allocs/op
BenchmarkTransfer500Mb/readBuffer_<_decrypted_plaintext-16     	       1	1773244421 ns/op	 295668442 bytes/sec	 8235792 B/op	  512078 allocs/op
BenchmarkTransfer500Mb/readBuffer_<_decrypted_plaintext-16     	       1	1766792935 ns/op	 296746149 bytes/sec	 8232872 B/op	  512089 allocs/op
BenchmarkTransfer500Mb/readBuffer_<_decrypted_plaintext-16     	       1	1770210938 ns/op	 296174536 bytes/sec	 8236392 B/op	  512085 allocs/op
BenchmarkTransfer500Mb/readBuffer_<_decrypted_plaintext-16     	       1	1774745574 ns/op	 295416415 bytes/sec	 8229096 B/op	  512089 allocs/op
BenchmarkTransfer500Mb/readBuffer_<_decrypted_plaintext-16     	       1	1764040402 ns/op	 297210707 bytes/sec	 8237704 B/op	  512098 allocs/op
PASS
ok  	github.com/libp2p/go-libp2p-noise	68.117s
```

## after

```
goos: darwin
goarch: amd64
pkg: github.com/libp2p/go-libp2p-noise
BenchmarkTransfer1MB/readBuffer_<_decrypted_plaintext-16         	     646	   1791000 ns/op	 586138181 bytes/sec	    6164 B/op	     147 allocs/op
BenchmarkTransfer1MB/readBuffer_<_decrypted_plaintext-16         	     652	   1766697 ns/op	 594118944 bytes/sec	    5969 B/op	     146 allocs/op
BenchmarkTransfer1MB/readBuffer_<_decrypted_plaintext-16         	     622	   1792639 ns/op	 585564766 bytes/sec	    5208 B/op	     148 allocs/op
BenchmarkTransfer1MB/readBuffer_<_decrypted_plaintext-16         	     632	   1773967 ns/op	 591679605 bytes/sec	    5449 B/op	     148 allocs/op
BenchmarkTransfer1MB/readBuffer_<_decrypted_plaintext-16         	     645	   1802342 ns/op	 582340372 bytes/sec	    5846 B/op	     147 allocs/op
BenchmarkTransfer1MB/readBuffer_>_encrypted_message-16           	     661	   1787814 ns/op	 587174250 bytes/sec	    5599 B/op	     145 allocs/op
BenchmarkTransfer1MB/readBuffer_>_encrypted_message-16           	     639	   1780549 ns/op	 589485802 bytes/sec	    5308 B/op	     147 allocs/op
BenchmarkTransfer1MB/readBuffer_>_encrypted_message-16           	     650	   1762170 ns/op	 595658443 bytes/sec	    5540 B/op	     146 allocs/op
BenchmarkTransfer1MB/readBuffer_>_encrypted_message-16           	     626	   1782546 ns/op	 588795887 bytes/sec	    5534 B/op	     148 allocs/op
BenchmarkTransfer1MB/readBuffer_>_encrypted_message-16           	     634	   1790120 ns/op	 586349625 bytes/sec	    5351 B/op	     148 allocs/op
BenchmarkTransfer100MB/readBuffer_>_encrypted_message-16         	       4	 265721914 ns/op	 394622404 bytes/sec	 1062310 B/op	   65733 allocs/op
BenchmarkTransfer100MB/readBuffer_>_encrypted_message-16         	       4	 267675429 ns/op	 391741711 bytes/sec	 1062242 B/op	   65732 allocs/op
BenchmarkTransfer100MB/readBuffer_>_encrypted_message-16         	       4	 268091615 ns/op	 391133483 bytes/sec	 1064282 B/op	   65732 allocs/op
BenchmarkTransfer100MB/readBuffer_>_encrypted_message-16         	       4	 269522798 ns/op	 389056060 bytes/sec	 1062270 B/op	   65732 allocs/op
BenchmarkTransfer100MB/readBuffer_>_encrypted_message-16         	       4	 266841962 ns/op	 392965180 bytes/sec	 1061186 B/op	   65731 allocs/op
BenchmarkTransfer100MB/readBuffer_<_decrypted_plaintext-16       	       4	 276498162 ns/op	 379240841 bytes/sec	 1075444 B/op	   65745 allocs/op
BenchmarkTransfer100MB/readBuffer_<_decrypted_plaintext-16       	       4	 278170864 ns/op	 376960554 bytes/sec	 1073394 B/op	   65745 allocs/op
BenchmarkTransfer100MB/readBuffer_<_decrypted_plaintext-16       	       4	 274700420 ns/op	 381722990 bytes/sec	 1071058 B/op	   65741 allocs/op
BenchmarkTransfer100MB/readBuffer_<_decrypted_plaintext-16       	       4	 273919183 ns/op	 382811662 bytes/sec	 1076502 B/op	   65746 allocs/op
BenchmarkTransfer100MB/readBuffer_<_decrypted_plaintext-16       	       4	 270810910 ns/op	 387205651 bytes/sec	 1078434 B/op	   65745 allocs/op
BenchmarkTransfer500Mb/readBuffer_>_encrypted_message-16         	       1	1543193839 ns/op	 339743040 bytes/sec	 8204040 B/op	  512032 allocs/op
BenchmarkTransfer500Mb/readBuffer_>_encrypted_message-16         	       1	1560171088 ns/op	 336045905 bytes/sec	 8212688 B/op	  512047 allocs/op
BenchmarkTransfer500Mb/readBuffer_>_encrypted_message-16         	       1	1541625046 ns/op	 340088697 bytes/sec	 8204072 B/op	  512033 allocs/op
BenchmarkTransfer500Mb/readBuffer_>_encrypted_message-16         	       1	1554192769 ns/op	 337338561 bytes/sec	 8204040 B/op	  512032 allocs/op
BenchmarkTransfer500Mb/readBuffer_>_encrypted_message-16         	       1	1539846834 ns/op	 340481345 bytes/sec	 8204152 B/op	  512035 allocs/op
BenchmarkTransfer500Mb/readBuffer_<_decrypted_plaintext-16       	       1	1573050012 ns/op	 333300576 bytes/sec	 8240776 B/op	  512088 allocs/op
BenchmarkTransfer500Mb/readBuffer_<_decrypted_plaintext-16       	       1	1586226384 ns/op	 330531189 bytes/sec	 8258792 B/op	  512117 allocs/op
BenchmarkTransfer500Mb/readBuffer_<_decrypted_plaintext-16       	       1	1590244124 ns/op	 329690957 bytes/sec	 8228352 B/op	  512085 allocs/op
BenchmarkTransfer500Mb/readBuffer_<_decrypted_plaintext-16       	       1	1588759401 ns/op	 329999165 bytes/sec	 8236360 B/op	  512083 allocs/op
BenchmarkTransfer500Mb/readBuffer_<_decrypted_plaintext-16       	       1	1601856651 ns/op	 327300978 bytes/sec	 8240424 B/op	  512084 allocs/op
PASS
ok  	github.com/libp2p/go-libp2p-noise	58.701s
```

## benchstat

```
name                                               old time/op    new time/op    delta
Transfer1MB/readBuffer_>_encrypted_message-16        1.79ms ± 1%    1.78ms ± 1%     ~     (p=0.690 n=5+5)
Transfer1MB/readBuffer_<_decrypted_plaintext-16      1.79ms ± 1%    1.79ms ± 1%     ~     (p=0.841 n=5+5)
Transfer100MB/readBuffer_<_decrypted_plaintext-16     289ms ± 1%     275ms ± 1%   -5.06%  (p=0.008 n=5+5)
Transfer100MB/readBuffer_>_encrypted_message-16       281ms ± 0%     268ms ± 1%   -4.94%  (p=0.008 n=5+5)
Transfer500Mb/readBuffer_>_encrypted_message-16       1.74s ± 1%     1.55s ± 1%  -11.16%  (p=0.008 n=5+5)
Transfer500Mb/readBuffer_<_decrypted_plaintext-16     1.77s ± 0%     1.59s ± 1%  -10.27%  (p=0.008 n=5+5)

name                                               old bytes/sec  new bytes/sec  delta
Transfer1MB/readBuffer_>_encrypted_message-16          587M ± 1%      589M ± 1%     ~     (p=0.690 n=5+5)
Transfer1MB/readBuffer_<_decrypted_plaintext-16        587M ± 1%      588M ± 1%     ~     (p=0.841 n=5+5)
Transfer100MB/readBuffer_<_decrypted_plaintext-16      362M ± 1%      382M ± 1%   +5.34%  (p=0.008 n=5+5)
Transfer100MB/readBuffer_>_encrypted_message-16        373M ± 0%      392M ± 1%   +5.19%  (p=0.008 n=5+5)
Transfer500Mb/readBuffer_>_encrypted_message-16        301M ± 1%      339M ± 1%  +12.56%  (p=0.008 n=5+5)
Transfer500Mb/readBuffer_<_decrypted_plaintext-16      296M ± 0%      330M ± 1%  +11.45%  (p=0.008 n=5+5)

name                                               old alloc/op   new alloc/op   delta
Transfer1MB/readBuffer_>_encrypted_message-16        5.05kB ± 4%    5.47kB ± 3%   +8.18%  (p=0.008 n=5+5)
Transfer1MB/readBuffer_<_decrypted_plaintext-16      5.11kB ± 5%    5.73kB ± 9%  +12.07%  (p=0.016 n=5+5)
Transfer100MB/readBuffer_<_decrypted_plaintext-16    1.07MB ± 0%    1.07MB ± 0%     ~     (p=0.222 n=5+5)
Transfer100MB/readBuffer_>_encrypted_message-16      1.06MB ± 0%    1.06MB ± 0%     ~     (p=0.222 n=5+5)
Transfer500Mb/readBuffer_>_encrypted_message-16      8.21MB ± 0%    8.20MB ± 0%   -0.06%  (p=0.032 n=5+4)
Transfer500Mb/readBuffer_<_decrypted_plaintext-16    8.23MB ± 0%    8.24MB ± 0%     ~     (p=0.310 n=5+5)

name                                               old allocs/op  new allocs/op  delta
Transfer1MB/readBuffer_>_encrypted_message-16           147 ± 1%       147 ± 1%     ~     (p=1.000 n=5+5)
Transfer1MB/readBuffer_<_decrypted_plaintext-16         146 ± 2%       147 ± 1%     ~     (p=0.532 n=5+5)
Transfer100MB/readBuffer_<_decrypted_plaintext-16     65.7k ± 0%     65.7k ± 0%     ~     (p=0.952 n=5+4)
Transfer100MB/readBuffer_>_encrypted_message-16       65.7k ± 0%     65.7k ± 0%     ~     (p=0.952 n=5+5)
Transfer500Mb/readBuffer_>_encrypted_message-16        512k ± 0%      512k ± 0%     ~     (p=0.119 n=5+5)
Transfer500Mb/readBuffer_<_decrypted_plaintext-16      512k ± 0%      512k ± 0%     ~     (p=0.714 n=5+5)
```