@Stebalien 2020-10-06
We intentionally _didn't_ do this in secio as it locks up a full 4k buffer while waiting for data when idle. However, that was _mostly_ an issue with streams because we have multiple streams per peer. For actual connections, this may be less of an issue 4k * 10k connections = 30MiB.

Ideally, we'd be able to ramp up/down:

* Start with a small prefix buffer.
* After reading two consecutive messages, pass in a larger buffer.
* After some idle time, interrupt the read (with a deadline... not sure if we can safely do this?) then swap out a small "idle" buffer.

But this is difficult in practice.

> Ideally, go-buffer-pool would support bufio.Reader backed by the buffer pool, but we'd need to clone the code for bufio.Reader, because one cannot pass the buffer.

This likely wouldn't help _too_ much as we need to hang on to the buffer for the lifetime of the connection.