@raulk 2020-10-07
> We intentionally didn't do this in secio as it locks up a full 4k buffer while waiting for data when idle. However, that was mostly an issue with streams because we have multiple streams per peer. 

But SecIO doesn't care about streams, no? For multiplexers, that can be a problem if it's added to the frontside (streams, 1 buffer per stream), but not to the backside (underlying connection, 1 buffer per connection).

The goal of this is to cushion the amount of syscalls we make when the channel is fast (i.e. there's data sitting in the receive buffer). Anything that relates to streams is already a layer above the syscall-performing layer.

In short, I think bufio only makes sense in secure channels (SecIO, Noise, TLS). Layers above that are already in userspace, and while bufio might speed up things, the tradeoff between memory and performance is more questionable.