@Stebalien 2020-10-07
Sorry, confusing sentence. We didn't do this with secio, to avoid locking up a 4k buffer per-connection (as far as I know). However, this is only _really_ an issue with idle streams in stream multiplexers, so it should be fine to do it in security transports.

I agree this makes sense to buffer each connection. Really, we might even want to do it one layer lower inside the TCP transport itself.